Remove the call to rack_builder ain #app_class and update README.md
The original Rack::Flash (http://gitrdoc.com/nakajima/rack-flash/tree/master)
had some issues which was fixed by Treeder
(https://github.com/treeder/rack-flash).
+
However, when *not* using Sinatra, the code still won't work as the call to
rack_builder is still present in the code, but when using with Sinatra it will
never be executed.

The offending call was removed.